+++ routes/tournaments.js
@@ -55,15 +55,18 @@ router.get('/current', function (req, res, next) {
 router.get('/admin', function (req, res, next) {
     axios.all([
         axios.get(`${req.app.locals.kcapp.api}/tournament/groups`),
+        axios.get(`${req.app.locals.kcapp.api}/tournament/preset`),
         axios.get(`${req.app.locals.kcapp.api}/player`),
         axios.get(`${req.app.locals.kcapp.api}/office`),
         axios.get(`${req.app.locals.kcapp.api}/venue`),
         axios.get(`${req.app.locals.kcapp.api}/match/modes`),
         axios.get(`${req.app.locals.kcapp.api}/match/types`),
-    ]).then(axios.spread((groups, players, offices, venues, modes, types) => {
+    ]).then(axios.spread((groups, presets, playersData, offices, venues, modes, types) => {
+        const players = _.reject(playersData.data, (player) => player.is_placeholder);
         res.marko(tournamentsAdminTemplate, {
             groups: groups.data,
-            players: players.data,
+            presets: presets.data,
+            players: _.sortBy(players, 'name'),
             offices: offices.data,
             venues: venues.data,
             modes: modes.data,
@@ -86,19 +89,18 @@ router.get('/:id/admin', function (req, res, next) {
         axios.get(`${req.app.locals.kcapp.api}/match/modes`),
         axios.get(`${req.app.locals.kcapp.api}/venue`),
     ]).then(axios.spread((playersResponse, tournamentResponse, matchesData, metadataResponse, groups, modes, venue) => {
-        var matches = matchesData.data;
-        var metadata = _.sortBy(metadataResponse.data, 'order_of_play');
-        var players = playersResponse.data;
-        var tournament = tournamentResponse.data;
+        const matches = matchesData.data;
+        const metadata = _.sortBy(metadataResponse.data, 'order_of_play');
+        const players = playersResponse.data;
+        const tournament = tournamentResponse.data;
 
-        var matchesMap = {};
-        for (var key in matches) {
+        const matchesMap = {};
+        for (let key in matches) {
             _.extend(matchesMap, _.object(_.map(matches[key], function (match) {
                 return [match.id, match]
             })));
         }
-
-        bracket.generateNew(metadata, matches, players, '', (err, brackets) => {
+        bracket.generate(tournament, metadata, matches, players, '', (err, brackets) => {
             res.marko(tournamentAdminTemplate, {
                 brackets: brackets,
                 tournament: tournament,
@@ -193,6 +195,67 @@ router.post('/admin', function (req, res, next) {
         });
 });
 
+/* Generate new tournament  */
+router.post('/admin/generate', function (req, res, next) {
+    const body = req.body;
+
+    axios.get(`${req.app.locals.kcapp.api}/tournament/preset/${body.preset_id}`)
+        .then(response => {
+            const preset = response.data;
+
+            const players = [];
+            for (let i = 0; i < body.group1.length; i++) {
+                const player = body.group1[i];
+                players.push({
+                    player_id: player.id,
+                    tournament_group_id: preset.group1_tournament_group_id.id
+                });
+            }
+            for (let i = 0; i < body.group2.length; i++) {
+                const player = body.group2[i];
+                players.push({
+                    player_id: player.id,
+                    tournament_group_id: preset.group2_tournament_group_id.id
+                });
+            }
+
+            const name = `${preset.name} ${moment().format('Do MMM')}`;
+            const shortName = `${name.substring(0, 1)}${moment().format('DDMM')}`;
+            const tournamentBody = {
+                name: name,
+                short_name: shortName,
+                is_playoffs: false,
+                players: players,
+                preset_id: body.preset_id,
+                manual_admin: true,
+                office_id: req.body.office_id
+            };
+            axios.post(`${req.app.locals.kcapp.api}/tournament/generate`, tournamentBody)
+                .then(response => {
+                    const tournament = response.data;
+                    res.send(tournament);
+                }).catch(error => {
+                    debug(`Error when generating new tournament: ${error}`);
+                    next(error);
+                });
+        }).catch(error => {
+            debug(`Error when generating new tournament: ${error}`);
+            next(error);
+        });
+});
+
+/* Generate playoffs tournament  */
+router.post('/admin/generate/playoffs/:id', function (req, res, next) {
+    axios.post(`${req.app.locals.kcapp.api}/tournament/generate/playoffs/${req.params.id}`)
+        .then(response => {
+            const tournament = response.data;
+            res.send(tournament);
+        }).catch(error => {
+            debug(`Error when generating playoffs tournament: ${error}`);
+            next(error);
+        });
+});
+
 /* Create new tournament group  */
 router.post('/admin/groups', function (req, res, next) {
     const body = req.body;
@@ -216,7 +279,7 @@ router.get('/:id', function (req, res, next) {
         axios.get(`${req.app.locals.kcapp.api}/tournament/${req.params.id}/statistics`),
         axios.get(`${req.app.locals.kcapp.api}/tournament/${req.params.id}/metadata`)
     ]).then(axios.spread((playersResponse, tournamentResponse, overviewData, matchesData, statisticsResponse, metadataResponse) => {
-        var statistics = statisticsResponse.data;
+        const statistics = statisticsResponse.data;
         if (!_.isEmpty(statistics)) {
             statistics.checkout_highest = _.sortBy(statistics.checkout_highest, (stats) => -stats.value);
             statistics.best_three_dart_avg = _.sortBy(statistics.best_three_dart_avg, (stats) => -stats.value);
@@ -229,12 +292,12 @@ router.get('/:id', function (req, res, next) {
             }
         }
 
-        var matches = matchesData.data;
-        var metadata = _.sortBy(metadataResponse.data, 'order_of_play');
-        var players = playersResponse.data;
-        var tournament = tournamentResponse.data;
+        const matches = matchesData.data;
+        const metadata = _.sortBy(metadataResponse.data, 'order_of_play');
+        const players = playersResponse.data;
+        const tournament = tournamentResponse.data;
 
-        var overview = overviewData.data;
+        const overview = overviewData.data;
         sortTournamentOverview(overview);
 
         if (tournament.playoffs_tournament_id !== null) {
@@ -242,7 +305,7 @@ router.get('/:id', function (req, res, next) {
                 axios.get(`${req.app.locals.kcapp.api}/tournament/${tournament.playoffs_tournament_id}/matches`),
                 axios.get(`${req.app.locals.kcapp.api}/tournament/${tournament.playoffs_tournament_id}/metadata`)
             ]).then(axios.spread((matchesResponse, metadataResponse) => {
-                bracket.generateNew(metadataResponse.data, matchesResponse.data, players, '', (err, brackets) => {
+                bracket.generate(tournament, metadataResponse.data, matchesResponse.data, players, '', (err, brackets) => {
                     res.marko(tournamentTemplate, {
                         brackets: brackets,
                         tournament: tournament,
@@ -257,7 +320,7 @@ router.get('/:id', function (req, res, next) {
                 next(error);
             });
         } else {
-            bracket.generateNew(metadata, matches, players, '', (err, brackets) => {
+            bracket.generate(tournament, metadata, matches, players, '', (err, brackets) => {
                 res.marko(tournamentTemplate, {
                     brackets: brackets,
                     tournament: tournament,